+++ routing/osrm2feature_map.hpp
@@ -136,7 +136,7 @@ public:
 
   template <class ToDo> void ForEachFtSeg(TOsrmNodeId nodeId, ToDo toDo) const
   {
-    pair<size_t, size_t> r = GetSegmentsRange(nodeId);
+    pair<uint32_t, uint32_t> r = GetSegmentsRange(nodeId);
     while (r.first != r.second)
     {
       OsrmMappingTypes::FtSeg s(m_segments[r.first]);
@@ -164,9 +164,9 @@ public:
   /// @name For unit test purpose only.
   //@{
   /// @return STL-like range [s, e) of segments indexies for passed node.
-  pair<size_t, size_t> GetSegmentsRange(uint32_t nodeId) const;
+  pair<uint32_t, uint32_t> GetSegmentsRange(uint32_t nodeId) const;
   /// @return Node id for segment's index.
-  TOsrmNodeId GetNodeId(size_t segInd) const;
+  TOsrmNodeId GetNodeId(uint32_t segInd) const;
 
   size_t GetSegmentsCount() const { return static_cast<size_t>(m_segments.size()); }
   //@}
